dto/src/main/java/com/gmail/sebastian/pisarski/dto/SpittleDto.java
package com.gmail.sebastian.pisarski.dto;

import java.util.Date;

import com.gmail.sebastian.pisarski.entity.Spittle;

public class SpittleDto extends BaseDto<Spittle> {

	private Long id;

	private String message;

	private String title;

	private Date time;

	private Long createUserId;

	private Long editUserId;

	private Date updateTime;

	public SpittleDto(Spittle entity) {
		super(entity);
	}

	public Long getId() {
		return id;
	}

	public void setId(Long id) {
		this.id = id;
	}

	public String getMessage() {
		return message;
	}

	public void setMessage(String message) {
		this.message = message;
	}

	public String getTitle() {
		return title;
	}

	public void setTitle(String title) {
		this.title = title;
	}

	public Date getTime() {
		return time;
	}

	public void setTime(Date time) {
		this.time = time;
	}

	public Long getCreateUserId() {
		return createUserId;
	}

	public void setCreateUserId(Long createUserId) {
		this.createUserId = createUserId;
	}

	public Long getEditUserId() {
		return editUserId;
	}

	public void setEditUserId(Long editUserId) {
		this.editUserId = editUserId;
	}

	public Date getUpdateTime() {
		return updateTime;
	}

	public void setUpdateTime(Date updateTime) {
		this.updateTime = updateTime;
	}

}
